+++ Rotation/HandlingRotation/Screens/iPhone/Method1Autosize/AutosizeScreen.xib.cs
@@ -103,7 +103,7 @@ namespace HandlingRotation.Screens.iPhone.Method1Autosize {
 			}
 			
 			// set the cell properties
-			cell.TextLabel.Text = this.tableItems[(int)indexPath.Row];
+			cell.TextLabel.Text = this.tableItems[indexPath.Row];
 			
 			// return the cell
 			return cell;
